+++ src/macros.h
@@ -35,31 +35,31 @@ IN THE SOFTWARE.
                        " Delete it before allocating", #var, var);             \
         exit(1);                                                               \
     }                                                                          \
-    var = (type*)malloc(sizeof(type)*(size));                                  \
+    var = (type*)malloc(sizeof(type) * (size_t)(size));                         \
     /*initialize memory for small structures*/                                 \
-    if(size == 1) memset(var, 0, sizeof(type));                                \
+    if((size_t)size == 1) memset(var, 0, sizeof(type));                        \
     if(!var) {                                                                 \
         dprintf_memerr("cannot allocate %d bytes for %s",                      \
-                       sizeof(type)*(size), #var);                             \
+                       sizeof(type)*(size_t)(size), #var);                     \
         exit(1);                                                               \
     }                                                                          \
     dprintf_mem("Added: %s %d %s (%x) %d bytes\n",                             \
-                 __FILE__, __LINE__, #var, var, sizeof(type)*(size));          \
+                 __FILE__, __LINE__, #var, var, sizeof(type)*(size_t)(size));  \
 } while(0)
 
 #define new0(var, type, size) do {                                             \
     new(var, type, size);                                                      \
-    memset(var, 0, sizeof(type) * (size));                                     \
+    memset(var, 0, sizeof(type) * (size_t)(size));                             \
 } while(0)
 
 #define renew(var, type, size) do {                                            \
-    var = (type*)realloc(var, sizeof(type)*(size));                            \
+    var = (type*)realloc(var, sizeof(type)*(size_t)(size));                    \
     if(!var) {                                                                 \
-        dprintf_memerr("cannot re-allocate %d bytes for %s", sizeof(type)*(size), #var);\
+        dprintf_memerr("cannot re-allocate %d bytes for %s", sizeof(type)*(size_t)(size), #var);\
         exit(1);                                                               \
     }                                                                          \
     dprintf_mem("Reallocated: %s %d %s (%x) %d bytes\n",                       \
-                 __FILE__, __LINE__, #var, var, sizeof(type)*(size));          \
+                 __FILE__, __LINE__, #var, var, sizeof(type)*(size_t)(size));  \
 } while(0)
 
 #define delete(var) do {                                                       \
@@ -77,7 +77,7 @@ IN THE SOFTWARE.
 } while(0)
 
 #define fill(var, character, type, size) do {                                  \
-    memset(var, character, sizeof(type) * (size));                             \
+    memset(var, character, sizeof(type) * (size_t)(size));                     \
 } while(0)
 
 /*********************************TEXTURE OPS**********************************/
@@ -107,7 +107,7 @@ glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);                  \
 
 #define CHECK_OPTION(options, o) char o = 0; do {                              \
     typeof(options) _o = strstr(options, #o);                                  \
-    int l = strlen(#o);                                                        \
+    size_t l = strlen(#o);                                                     \
     /* check bounds */                                                         \
     if(NULL != _o && (' '==*(_o+l)||0==*(_o+l)) && (' '==*(_o-1)||0==*(_o-1))){\
         o = 1;                                                                 \
